Deduplicate list of statuses that count as 'transferred'
