+++ client/src/lesc/components/deflectionStatusChipUtils.js
@@ -1,22 +1,7 @@
 import { DateTime } from 'luxon';
 
 import { isValidDeflection, isValidIncident } from '../../utils/validators';
-
-export const CUSTODY_TRANSFERRED_SUBJECT_STATUSES = new Set([
-  'AWAITING_INTAKE',
-  'READY_FOR_INTAKE',
-  'FAILED_INTAKE',
-  'IN_MEDICAL_INTAKE',
-  'IN_CHAIR',
-  'RELEASED',
-  'EXITED',
-  'DEATH_IN_FACILITY',
-  'DEATH_IN_CUSTODY',
-]);
-
-export function isCustodyTransferredStatus (subjectStatus) {
-  return CUSTODY_TRANSFERRED_SUBJECT_STATUSES.has(subjectStatus);
-}
+import { isCustodyTransferredStatus } from './custodyTransferStatus';
 
 export function isExpiredBeforeTransfer (deflection, now) {
   if (!deflection || deflection.subjectStatus !== 'DETAINED') return false;
