+++ client/src/lesc/components/sfpdDocuments.js
@@ -1,15 +1,7 @@
+import { isCustodyTransferredStatus } from './custodyTransferStatus';
+
 export function getSfpdDocuments ({ deflection, view647fForm, download647fForm }) {
-  const isCustodyTransferred = [
-    'AWAITING_INTAKE',
-    'READY_FOR_INTAKE',
-    'FAILED_INTAKE',
-    'IN_MEDICAL_INTAKE',
-    'IN_CHAIR',
-    'RELEASED',
-    'EXITED',
-    'DEATH_IN_FACILITY',
-    'DEATH_IN_CUSTODY',
-  ].includes(deflection?.subjectStatus);
+  const isCustodyTransferred = isCustodyTransferredStatus(deflection?.subjectStatus);
   const doc647f = deflection?.deflectionDocuments?.find(d => d.formId === '647f');
 
   if (!isCustodyTransferred || !deflection?.transferredAt || !doc647f) {
